+++ package.json
@@ -28,11 +28,12 @@
     "@nestjs-query/query-typeorm": "^0.28.1",
     "@nestjs/common": "^8.0.6",
     "@nestjs/core": "^8.0.0",
+    "@nestjs/cqrs": "^8.0.0",
     "@nestjs/elasticsearch": "^8.0.0",
     "@nestjs/platform-express": "^8.0.0",
     "@nestjs/swagger": "^5.0.9",
     "@nestjs/typeorm": "^8.0.2",
-    "@polkadot/api": "^5.2.1",
+    "@polkadot/api": "^6.5.2",
     "axios": "^0.21.1",
     "cache-manager": "^3.4.4",
     "class-transformer": "^0.4.0",
